implement fixes
